Quito la actionbar de la actividad de registro, cambio un poco los layouts y comento el codigo que estaba haciendo de insercion de imágenes

+++ app/src/main/java/com/example/f1bets/ui/profile/UserFragment.kt
@@ -10,6 +10,7 @@ import com.example.f1bets.R
 import com.example.f1bets.StartActivity
 import com.example.f1bets.databinding.FragmentUserBinding
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import com.google.android.material.snackbar.Snackbar
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
@@ -43,26 +44,33 @@ class UserFragment : Fragment() {
 
 
 
+
     private fun userDelete(){
         val user = Firebase.auth.currentUser!!
-        user.delete().addOnCompleteListener { task ->
-            if (task.isSuccessful) {
-                val dialogBuilder = MaterialAlertDialogBuilder(binding.root.context)
-                    .setTitle("F1Bets")
-                    .setMessage(getString(R.string.txtDeleteAccount))
-                    .setPositiveButton(getString(R.string.txtContinue)) { dialog, _ ->
-                        val i = Intent(requireActivity(), StartActivity::class.java)
-                        startActivity(i)
-                        auth.signOut()
-                        requireActivity().finish()
-                        dialog.dismiss()
-                    }
-                    .setNegativeButton(getString(R.string.txtCancel)) { dialog, _ ->
-                        dialog.dismiss()
-                    }
-                dialogBuilder.show()
+        if (user != null) {
+            user.delete().addOnCompleteListener { task ->
+                if (task.isSuccessful) {
+                    val dialogBuilder = MaterialAlertDialogBuilder(binding.root.context)
+                        .setTitle(R.string.app_name)
+                        .setMessage(getString(R.string.txtDeleteAccount))
+                        .setPositiveButton(getString(R.string.txtContinue)) { dialog, _ ->
+                            val i = Intent(requireActivity(), StartActivity::class.java)
+                            startActivity(i)
+                            auth.signOut()
+                            requireActivity().finish()
+                            dialog.dismiss()
+                        }
+                        .setNegativeButton(getString(R.string.txtCancel)) { dialog, _ ->
+                            dialog.dismiss()
+                        }
+                    dialogBuilder.show()
+                } else {
+                    Snackbar.make(binding.root, R.string.errDeleteAccount, Snackbar.LENGTH_LONG).show()
+                }
             }
+        } else {
         }
+
     }
 
     private fun showConfirmationDialog() {
